+++ migrations/01_create_table_blacklist_domains.sql
@@ -1,0 +1,5 @@
+CREATE TABLE IF NOT EXISTS blacklist_domains
+(
+    id     SERIAL PRIMARY KEY,
+    domain VARCHAR(256) NOT NULL
+)